split up code to different utils
